Module Index

ModeCouplingTheory.EulerSolver

ModeCouplingTheory.ExponentiallyDecayingKernel

ModeCouplingTheory.FuchsSolver

ModeCouplingTheory.MCTProblem

ModeCouplingTheory.ModeCouplingKernel

ModeCouplingTheory.MultiComponentModeCouplingKernel

ModeCouplingTheory.SchematicDiagonalKernel

ModeCouplingTheory.SchematicF123Kernel

ModeCouplingTheory.SchematicF1Kernel

ModeCouplingTheory.SchematicF2Kernel

ModeCouplingTheory.SchematicMatrixKernel

ModeCouplingTheory.find_relaxation_time

ModeCouplingTheory.solve

ModeCouplingTheory.solve_steady_state

Detailed API

ModeCouplingTheory.EulerSolver — Method

EulerSolver(problem::MCTProblem; t_max=10.0^2, Δt=10^-3, verbose=false)

Constructs a solver object that, when called solve upon will solve an MCTProblem using a forward Euler method. It will discretise the integral using a Trapezoidal rule. Use this solver only for testing purposes. It is a wildy inefficient way to solve MCT-like equations.

Arguments:

problem an instance of MCTProblem

t_max when this time value is reached, the integration returns

Δt fixed time step

verbose if true, information will be printed to STDOUT

ModeCouplingTheory.ExponentiallyDecayingKernel — Type

ExponentiallyDecayingKernel{T<:Number} <: MemoryKernel

Scalar kernel with fields λ and τ which when called returns λ exp(-t/τ).

ModeCouplingTheory.FuchsSolver — Method

FuchsSolver(problem; N=32, Δt=10^-10, t_max=10.0^10, max_iterations=10^4, tolerance=10^-10, verbose=false)

Uses the algorithm devised by Fuchs et al. to solve the MCTProblem.

Arguments:

problem: an instance of MCTProblem

N: The number of time points in the interval is equal to 4N

t_max: when this time value is reached, the integration returns

Δt: starting time step, this will be doubled repeatedly

max_iterations: the maximal number of iterations before convergence is reached for each time doubling step

tolerance: while the error is bigger than this value, convergence is not reached. The error by default is computed as the absolute sum of squares

verbose: if true, information will be printed to STDOUT

ModeCouplingTheory.MCTProblem — Method

MCTProblem(α, β, γ, F₀::T, ∂ₜF₀::T, kernel::MemoryKernel) where T

Constructor of the MCTProblem type. It requires that F₀ and ∂ₜF₀ have the same type.

Arguments:

α: coefficient in front of the second derivative term. If α and F₀ are both vectors, α will automatically be converted to a diagonal matrix, to make them compatible.

β: coefficient in front of the first derivative term. If β and F₀ are both vectors, β will automatically be converted to a diagonal matrix, to make them compatible.

γ: coefficient in front of the second derivative term. If γ and F₀ are both vectors, γ will automatically be converted to a diagonal matrix, to make them compatible.

F₀: initial condition of F(t)

∂ₜF₀ initial condition of the derivative of F(t)

kernel instance of a MemoryKernel that when called on F₀ and t=0, evaluates to the initial condition of the memory kernel.

ModeCouplingTheory.ModeCouplingKernel — Method

ModeCouplingKernel(ρ, kBT, m, k_array, Sₖ)

Constructor of a ModeCouplingKernel. It implements the kernel K(k,t) = ρ kBT / (16π^3m) ∫dq V^2(k,q) F(q,t) F(k-q,t) in which k and q are vectors.

Arguments:

ρ: number density

kBT: Thermal energy

m : particle mass

k_array: vector of wavenumbers at which the structure factor is known

Sₖ: structure factor

Returns:

an instance k of ModeCouplingKernel <: MemoryKernel, which can be called both in-place and out-of-place: k(out, F, t) out = k(F, t)

ModeCouplingTheory.MultiComponentModeCouplingKernel — Method

MultiComponentModeCouplingKernel(ρ, kBT, m, k_array, Sₖ)

Constructor of a MultiComponentModeCouplingKernel. It implements the kernel Kαβ(k,t) = ρ / (2 xα xβ (2π)³) Σμνμ'ν' ∫dq Vμ'ν'(k,q) Fμμ'(q,t) Fνν'(k-q,t) Vμν(k,q) in which k and q are vectors and α and β species labels.

Arguments:

ρ: vector of number densities for each species

kBT: Thermal energy

m : vector of particle masses for each species

k_array: vector of wavenumbers at which the structure factor is known

Sₖ: a Vector of Nk SMatrixs containing the structure factor of each component at each wave number

Returns:

an instance k of ModeCouplingKernel <: MemoryKernel, which can be called both in-place and out-of-place: k = MultiComponentModeCouplingKernel(ρ, kBT, m, k_array, Sₖ) out = k(F, t) k(out, F, t)

ModeCouplingTheory.SchematicDiagonalKernel — Type

SchematicDiagonalKernel{T<:Union{SVector, Vector}} <: MemoryKernel

Matrix kernel with field λ which when called returns Diagonal(λ .* F .^ 2), i.e., it implements a non-coupled system of SchematicF2Kernels.

ModeCouplingTheory.SchematicF123Kernel — Type

SchematicF1Kernel{T<:Number} <: MemoryKernel

Scalar kernel with fields λ1, λ2, and λ3 which when called returns λ1 * F^1 + λ2 * F^2 + λ3 * F^3.

ModeCouplingTheory.SchematicF1Kernel — Type

SchematicF1Kernel{T<:Number} <: MemoryKernel

Scalar kernel with field λ which when called returns λ F.

ModeCouplingTheory.SchematicF2Kernel — Type

SchematicF2Kernel{T<:Number} <: MemoryKernel

Scalar kernel with field λ which when called returns λ F^2.

ModeCouplingTheory.SchematicMatrixKernel — Type

SchematicMatrixKernel{T<:Union{SVector, Vector}} <: MemoryKernel

Matrix kernel with field λ which when called returns λ * F * Fᵀ, i.e., it implements Kαβ = λFαFβ.

ModeCouplingTheory.find_relaxation_time — Method

find_relaxation_time(t, F; threshold=exp(-1), mode=:log)

Finds the time at which a signal first decreases below some threshold. Uses interpolation to find an accurate result

Arguments:

t: a vector of time values at which the signal F is known

F: the signal

threshold: see above

mode: uses interpolation either in a logarithmic or linear space.

ModeCouplingTheory.solve — Method

solve(problem::MCTProblem, solver::Solver, kernel::MemoryKernel)

Solves the MCTproblem with the provided kernel using solver. Search for a specific solver or kernel object to find more specific information.

Returns:

t an array of time values

F The solution in an array of which the last dimension corresponds to the time.

K The memory kernel corresponding to each F

ModeCouplingTheory.solve_steady_state — Method

solve_steady_state(γ, F₀, kernel; tolerance=10^-8, max_iterations=10^6, verbose=false)

Finds the steady-state solution (non-ergodicity parameter) of the generalized Langevin equation by recursive iteration of F = (K + γ)⁻¹ * K(F) * F₀

Arguments:

γ: parameter in front of the linear term in F

F₀: initial condition of F. This is also the initial condition of the rootfinding method.

kernel: callable memory kernel

max_iterations: the maximal number of iterations before convergence is reached

tolerance: while the error is bigger than this value, convergence is not reached. The error by default is computed as the absolute sum of squares between successive iterations

verbose: if true, information will be printed to STDOUT

Returns:

The steady state solution