Module Index

ModeCouplingTheory.BetaScalingEquation

ModeCouplingTheory.EulerSolver

ModeCouplingTheory.ExponentiallyDecayingKernel

ModeCouplingTheory.InterpolatingKernel

ModeCouplingTheory.MemoryEquation

ModeCouplingTheory.MemoryEquationCoefficients

ModeCouplingTheory.MemoryEquationSolution

ModeCouplingTheory.ModeCouplingKernel

ModeCouplingTheory.MultiComponentModeCouplingKernel

ModeCouplingTheory.SchematicDiagonalKernel

ModeCouplingTheory.SchematicF123Kernel

ModeCouplingTheory.SchematicF1Kernel

ModeCouplingTheory.SchematicF2Kernel

ModeCouplingTheory.SchematicMatrixKernel

ModeCouplingTheory.SjogrenKernel

ModeCouplingTheory.TaggedModeCouplingKernel

ModeCouplingTheory.TaggedMultiComponentModeCouplingKernel

ModeCouplingTheory.TaggedSchematicF2Kernel

ModeCouplingTheory.TimeDoublingSolver

ModeCouplingTheory.allocate_results!

ModeCouplingTheory.allocate_temporary_arrays

ModeCouplingTheory.convertresults

ModeCouplingTheory.do_time_steps!

ModeCouplingTheory.evaluate_kernel

ModeCouplingTheory.evaluate_kernel!

ModeCouplingTheory.find_error

ModeCouplingTheory.find_relaxation_time

ModeCouplingTheory.initialize_F_temp!

ModeCouplingTheory.initialize_F_temp!

ModeCouplingTheory.initialize_K_temp!

ModeCouplingTheory.initialize_integrals!

ModeCouplingTheory.initialize_integrals!

ModeCouplingTheory.initialize_output_arrays

ModeCouplingTheory.mymul!

ModeCouplingTheory.new_time_mapping!

ModeCouplingTheory.regula_falsi

ModeCouplingTheory.solve

ModeCouplingTheory.solve_steady_state

ModeCouplingTheory.update_F!

ModeCouplingTheory.update_Fuchs_parameters!

ModeCouplingTheory.update_Fuchs_parameters!

ModeCouplingTheory.update_K!

ModeCouplingTheory.update_integrals!

Detailed API

ModeCouplingTheory.BetaScalingEquation — Method

BetaScalingEquation(λ, σ, t₀, δ=0.0)

Defines the β-scaling equation of MCT, which is a scalar equation for a single scaling function g(t), determined by scalar parameters λ (the MCT exponent parameter, 1/2<=λ<1), the distance parameter to the glass transition σ, and (for convenience) an arbitrary time scale t₀ that just shifts the results. Optionally, a "hopping parameter" δ can be given (defaults to zero).

The MCT exponents a and b will be automatically calculated from λ.

ModeCouplingTheory.EulerSolver — Method

EulerSolver(equation::MemoryEquation; t_max=10.0^2, Δt=10^-3, verbose=false)

Constructs a solver object that, when called solve upon will solve an MemoryEquation using a forward Euler method. It will discretise the integral using a Trapezoidal rule. Use this solver only for testing purposes. It is a wildy inefficient way to solve MCT-like equations.

Arguments:

equation an instance of MemoryEquation

t_max when this time value is reached, the integration returns

Δt fixed time step

verbose if true, information will be printed to STDOUT

ModeCouplingTheory.ExponentiallyDecayingKernel — Type

ExponentiallyDecayingKernel{T<:Number} <: MemoryKernel

Scalar kernel with fields ν and τ which when called returns ν exp(-t/τ).

ModeCouplingTheory.InterpolatingKernel — Method

InterpolatingKernel(t, M; k=1)

Uses the package Dierckx to provide a kernel that interpolates the data M defined on grid t using spline interpolation of degree k.

ModeCouplingTheory.MemoryEquation — Method

MemoryEquation(α, β, γ, F₀::T, ∂ₜF₀::T, kernel::MemoryKernel) where T

Arguments:

α: coefficient in front of the second derivative term. If α and F₀ are both vectors, α will automatically be converted to a diagonal matrix, to make them compatible.

β: coefficient in front of the first derivative term. If β and F₀ are both vectors, β will automatically be converted to a diagonal matrix, to make them compatible.

γ: coefficient in front of the second derivative term. If γ and F₀ are both vectors, γ will automatically be converted to a diagonal matrix, to make them compatible.

δ: coefficient in front of the constant term. δ and F0 must by types that can be added together. If δ is a number and F0 is a vector, this conversion will be done automatically.

F₀: initial condition of F(t)

∂ₜF₀ initial condition of the derivative of F(t)

kernel instance of a MemoryKernel that when called on F₀ and t=0, evaluates to the initial condition of the memory kernel.

ModeCouplingTheory.MemoryEquationCoefficients — Method

MemoryEquationCoefficients(a, b, c, d, F₀)

Constructor of the MemoryEquationCoefficients struct, which holds the values of the coefficients α, β, γ and δ. It will convert common cases automatically to make the types compatible. For example, if α and F0 are both given as vectors, α is converted to a Diagonal matrix.

ModeCouplingTheory.MemoryEquationSolution — Type

MemoryEquationSolution

solution object that holds the solution of an AbstractMemoryEquation. It has 4 fields t: array of t values F: array of F for all t K: array of K for all T solver: solver object that holds the solver settings

an MCT object can be indexed such that sol=MemoryEquationSolution sol[2] gives the F[2] for all t.

ModeCouplingTheory.ModeCouplingKernel — Method

ModeCouplingKernel(ρ, kBT, m, k_array, Sₖ)

Constructor of a ModeCouplingKernel. It implements the kernel K(k,t) = ρ kBT / (16π^3m) ∫dq V^2(k,q) F(q,t) F(k-q,t) in which k and q are vectors.

Arguments:

ρ: number density

kBT: Thermal energy

m : particle mass

k_array: vector of wavenumbers at which the structure factor is known

Sₖ: structure factor

Returns:

an instance k of ModeCouplingKernel <: MemoryKernel, which can be called both in-place and out-of-place: k(out, F, t) out = k(F, t)

ModeCouplingTheory.MultiComponentModeCouplingKernel — Method

MultiComponentModeCouplingKernel(ρ, kBT, m, k_array, Sₖ)

Constructor of a MultiComponentModeCouplingKernel. It implements the kernel Kαβ(k,t) = ρ / (2 xα xβ (2π)³) Σμνμ'ν' ∫dq Vμ'ν'(k,q) Fμμ'(q,t) Fνν'(k-q,t) Vμν(k,q) in which k and q are vectors and α and β species labels.

Arguments:

ρ: vector of number densities for each species

kBT: Thermal energy

m : vector of particle masses for each species

k_array: vector of wavenumbers at which the structure factor is known

Sₖ: a Vector of Nk SMatrixs containing the structure factor of each component at each wave number

Returns:

an instance k of ModeCouplingKernel <: MemoryKernel, which can be called both in-place and out-of-place: k = MultiComponentModeCouplingKernel(ρ, kBT, m, k_array, Sₖ) out = k(F, t) k(out, F, t)

ModeCouplingTheory.SchematicDiagonalKernel — Type

SchematicDiagonalKernel{T<:Union{SVector, Vector}} <: MemoryKernel

Matrix kernel with field ν which when called returns Diagonal(ν .* F .^ 2), i.e., it implements a non-coupled system of SchematicF2Kernels.

ModeCouplingTheory.SchematicF123Kernel — Type

SchematicF1Kernel{T<:Number} <: MemoryKernel

Scalar kernel with fields ν1, ν2, and ν3 which when called returns ν1 * F^1 + ν2 * F^2 + ν3 * F^3.

ModeCouplingTheory.SchematicF1Kernel — Type

SchematicF1Kernel{T<:Number} <: MemoryKernel

Scalar kernel with field ν which when called returns ν F.

ModeCouplingTheory.SchematicF2Kernel — Type

SchematicF2Kernel{T<:Number} <: MemoryKernel

Scalar kernel with field ν which when called returns ν F^2.

ModeCouplingTheory.SchematicMatrixKernel — Type

SchematicMatrixKernel{T<:Union{SVector, Vector}} <: MemoryKernel

Matrix kernel with field ν which when called returns ν * F * Fᵀ, i.e., it implements Kαβ = νFαFβ.

ModeCouplingTheory.SjogrenKernel — Type

SjogrenKernel

Memory kernel that implements the kernel K[1] = ν1 F[1]^2, K[2] = ν2 F[1] F[2]. Consider using Static Vectors for performance.

ModeCouplingTheory.TaggedModeCouplingKernel — Method

TaggedModeCouplingKernel(ρ, kBT, m, k_array, Sₖ, sol)

Constructor of a Tagged ModeCouplingKernel. It implements the kernel K(k,t) = ρ kBT / (8π^3m) ∫dq V^2(k,q) F(q,t) Fs(k-q,t) in which k and q are vectors.

Arguments:

ρ: number density

kBT: Thermal energy

m : particle mass

k_array: vector of wavenumbers at which the structure factor is known

Sₖ: vector with the elements of the structure factor

sol: a solution object of an equation with a ModeCouplingKernel.

Returns:

an instance k of TaggedModeCouplingKernel <: MemoryKernel, which can be called both in-place and out-of-place: k(out, F, t) out = k(F, t)

ModeCouplingTheory.TaggedMultiComponentModeCouplingKernel — Method

TaggedMultiComponentModeCouplingKernel(ρ, kBT, m, k_array, Sₖ, sol)

Constructor of a Tagged ModeCouplingKernel. It implements the kernel K(k,t) = ρ kBT / (8π^3 mₛ) Σαβ ∫dq V^2sαβ(k,q) Fαβ(q,t) Fₛ(k-q,t) in which k and q are vectors.

Arguments:

s: index of the sepcies to tag

ρ: number density

kBT: Thermal energy

m : particle mass

k_array: vector of wavenumbers at which the structure factor is known

Sₖ: vector with the elements of the structure factor

sol: a solution object of an equation with a MultiComponentModeCouplingKernel.

Returns:

an instance k of TaggedMultiComponentModeCouplingKernel <: MemoryKernel, which can be called both in-place and out-of-place: k(out, F, t) out = k(F, t)

ModeCouplingTheory.TaggedSchematicF2Kernel — Type

SchematicF2Kernel{T<:Number} <: MemoryKernel

Scalar tagged particle kernel with field ν which when called returns ν F*Fs.

ModeCouplingTheory.TimeDoublingSolver — Method

TimeDoublingSolver(N=32, Δt=10^-10, t_max=10.0^10, max_iterations=10^4, tolerance=10^-10, verbose=false, ismutable=true)

Uses the algorithm devised by Fuchs et al.

Arguments:

equation: an instance of MemoryEquation

N: The number of time points in the interval is equal to 4N

t_max: when this time value is reached, the integration returns

Δt: starting time step, this will be doubled repeatedly

max_iterations: the maximal number of iterations before convergence is reached for each time doubling step

tolerance: while the error is bigger than this value, convergence is not reached. The error by default is computed as the absolute sum of squares

verbose: if true, information will be printed to STDOUT

ismutable: if true and if the type of F is mutable, the solver will try to avoid allocating many temporaries

ModeCouplingTheory.allocate_results! — Method

allocate_results!(t_array, F_array, K_array, equation::AbstractMemoryEquation, solver::TimeDoublingSolver, temp_arrays::SolverCache; istart=2(solver.N)+1, iend=4(solver.N))

pushes the found solution, stored in temp_arrays with indices istart until istop to the output arrays.

ModeCouplingTheory.allocate_temporary_arrays — Method

allocate_temporary_arrays(equation::AbstractMemoryEquation, solver::TimeDoublingSolver)

Returns a SolverCache containing several arrays that are used for intermediate calculations.

ModeCouplingTheory.convertresults — Method

convertresults(F_array::Vector{T}, K_array::Vector{Diagonal{T, T2}}) where {T2, T}

converts the arrays of arrays into multidimensional arrays for ease of use.

ModeCouplingTheory.do_time_steps! — Method

do_time_steps!(equation::MemoryEquation, solver::TimeDoublingSolver, kernel::MemoryKernel, temp_arrays::SolverCache)

Solves the equation on the time points with index 2N+1 until 4N, for each point doing a recursive iteration to find the solution to the nonlinear equation C1 F = -C2 M(F) + C3.

ModeCouplingTheory.evaluate_kernel! — Method

evaluate_kernel!(out, kernel::MemoryKernel, F, t)

Evaluates the memory kernel in-place, overwriting the elements of the out variable. It may mutate the content of kernel.

ModeCouplingTheory.evaluate_kernel — Method

evaluate_kernel(kernel::MemoryKernel, F, t)

Evaluates the memory kernel out-place. It may mutate the content of kernel.

Returns

out the kernel evaluated at (F, t)

ModeCouplingTheory.find_error — Method

find_error(F_new::T, F_old::T) where T

Finds the error between a new and old iteration of F. The returned scalar will be compared to the tolerance to establish convergence.

ModeCouplingTheory.find_relaxation_time — Method

find_relaxation_time(t, F; threshold=exp(-1), mode=:log)

Finds the time at which a signal first decreases below some threshold. Uses interpolation to find an accurate result. if it never decreases below the threshold, returns Inf.

Arguments:

t: a vector of time values at which the signal F is known

F: the signal

threshold: see above

mode: uses interpolation either in a logarithmic or linear space.

ModeCouplingTheory.initialize_F_temp! — Method

initialize_F_temp!(equation::BetaScalingEquation, solver::TimeDoublingSolver, temp_arrays::SolverCache)

Fills the first 2N entries of the temporary arrays needed for solving the β-scaling equation with a an adapted Fuchs scheme.

In particular, this initializes g(t) = (t/t₀)^-a.

ModeCouplingTheory.initialize_F_temp! — Method

initialize_F_temp!(equation::MemoryEquation, solver::TimeDoublingSolver, temp_arrays::SolverCache)

Fills the first 2N entries of the temporary arrays of F using forward Euler without a memory kernel in order to kickstart the algorithm' scheme.

ModeCouplingTheory.initialize_K_temp! — Method

initialize_K_temp!(solver::TimeDoublingSolver, kernel::MemoryKernel, temp_arrays::SolverCache)

Evaluates the memory kernel at the first 2N time points.

ModeCouplingTheory.initialize_integrals! — Method

initialize_integrals!(equation::AbstractMemoryEquation, solver::TimeDoublingSolver, temp_arrays::SolverCache)

Initializes the integrals on the first 2N time points as prescribed in the literature.

ModeCouplingTheory.initialize_integrals! — Method

initialize_integrals!(equation::BetaScalingEquation, solver::TimeDoublingSolver, temp_arrays::SolverCache)

Initializes the integrals over the first 2N time points for the solution of the β-scaling equation, using the known critical decay law as the short-time asymptote.

ModeCouplingTheory.initialize_output_arrays — Method

initialize_output_arrays(equation::AbstractMemoryEquation)

initializes arrays that the solver will push results into.

ModeCouplingTheory.mymul! — Method

" mymul!(c,a,b,α,β)

prescribes how types used in this solver should be multiplied in place. In particular, it performs C.= βC .+ αa*b. defaults to mul!(c,a,b,α,β)

ModeCouplingTheory.new_time_mapping! — Method

newtimemapping!(equation::AbstractMemoryEquation, solver::TimeDoublingSolver, temp_arrays::SolverCache)

Performs the time mapping central to Fuchs' algorithm with the conventions prescribed in "Flenner, Elijah, and Grzegorz Szamel. Physical Review E 72.3 (2005): 031508". Performs them inplace if solver.inplace = true in order to avoid unnecessary allocations.

ModeCouplingTheory.regula_falsi — Method

regula_falsi(x0::Float64, x1::Float64, f; accuracy=10^-10, max_iterations=10^4)

Implements a simple "regular falsi" search to find a point x where f(x)=0 within the given accuracy, starting in the interval [x0,x1]. For purists, f(x) should have exactly one root in the given interval, and the scheme will then converge to that. The present code is a bit more robust, at the expense of not guaranteeing convergence strictly.

Returns a point x such that approximately f(x)==0 (unless the maximum number of iterations has been exceeded). If exactly one of the roots of f(x) lie in the initially given interval [x0,x1], the returned x will be the approximation to that root.

Arguments:

x0: lower bound of the initial search interval

x1: upper bound of the initial search interval

f: a function accepting one real value, return one real value

ModeCouplingTheory.solve — Method

solve(equation::AbstractMemoryEquation, solver::Solver)

Solves the AbstractMemoryEquation with the provided kernel using solver. Search for a specific solver or kernel object to find more specific information.

If no solver is provided, it uses the default TimeDoublingSolver.

Returns:

t an array of time values

F The solution in an array of which the last dimension corresponds to the time.

K The memory kernel corresponding to each F

ModeCouplingTheory.solve_steady_state — Method

solve_steady_state(γ, F₀, kernel; tolerance=10^-8, max_iterations=10^6, verbose=false, inplace=true)

Finds the steady-state solution (non-ergodicity parameter) of the generalized Langevin equation by recursive iteration of F = (K + γ)⁻¹ * K(F) * F₀ This function assumes δ = 0, since that would lead to a divergence.

Arguments:

γ: parameter in front of the linear term in F

F₀: initial condition of F. This is also the initial condition of the rootfinding method.

kernel: callable memory kernel

max_iterations: the maximal number of iterations before convergence is reached

tolerance: while the error is bigger than this value, convergence is not reached. The error by default is computed as the absolute sum of squares between successive iterations

verbose: if true, information will be printed to STDOUT

inplace: if true and if the type of F is mutable, the solver will try to avoid allocating many temporaries. default =true``

Returns:

The steady state solution

ModeCouplingTheory.update_F! — Method

update_F!(solver::TimeDoublingSolver, temp_arrays::SolverCache, it::Int)

updates F using the formula c1F = -KC2 + C3.

ModeCouplingTheory.update_Fuchs_parameters! — Method

update_Fuchs_parameters!(equation::BetaScalingEquation, solver::TimeDoublingSolver, temp_arrays::SolverCache, it::Int)

Updates the parameters that are needed to solve the β-scaling equation numerically with Fuchs' scheme.

ModeCouplingTheory.update_Fuchs_parameters! — Method

update_Fuchs_parameters!(equation::MemoryEquation, solver::TimeDoublingSolver, temp_arrays::SolverCache, it::Int)

Updates the parameters c1, c2, c3, according to the appendix of "Flenner, Elijah, and Grzegorz Szamel. Physical Review E 72.3 (2005): 031508" using the naming conventions from that paper. If F is mutable (and therefore also c1,c2,c3), it will update the variables in place, otherwise it will create new copies. This is controlled by the solver.inplace setting.

ModeCouplingTheory.update_K! — Method

update_K!(solver::TimeDoublingSolver, kernel::MemoryKernel, temp_arrays::SolverCache, it::Int)

evaluates the memory kernel, updating the value in solver.temparrays.Ktemp

ModeCouplingTheory.update_integrals! — Method

update_integrals!(solver::TimeDoublingSolver, equation::AbstractMemoryEquation, temp_arrays::SolverCache, it::Int)

Update the discretisation of the integral of F and K, see the literature for details.